+++ tests/testthat/test-templates.R
@@ -204,6 +204,34 @@ test_that("eta_fu_template works as expected for 2019 data", {
 })
 
 
+test_that("eta_fu_template() works with tidy fu allocation data", {
+  tidy_specified_iea_data <- load_tidy_iea_df() %>% 
+    specify_all()
+  Eta_fu_template_2019 <- load_fu_allocation_data(sample_fu_allocation_table_path(2019)) %>% 
+    tidy_fu_allocation_table() %>% 
+    eta_fu_template(tidy_specified_iea_data = tidy_specified_iea_data)
+  
+  # Now try a couple tests.
+  # These tests are same as the tests in the previous test function.
+  expect_equal(Eta_fu_template_2019$Machine[[1]], "Automobiles")
+  expect_equal(Eta_fu_template_2019$Machine[[nrow(Eta_fu_template_2019)]], "Non-energy")
+  expect_equal(as.character(Eta_fu_template_2019$Quantity[[1]]), "E.dot_machine")
+  expect_equal(as.character(Eta_fu_template_2019$Quantity[[nrow(Eta_fu_template_2019)]]), "phi.u")
+  
+  eu_products <- Eta_fu_template_2019$Eu.product %>% unique() %>% as.character()
+  # Check that the order is as expected.
+  expect_equivalent(eu_products, c("MD", "Light", "HTH.600.C", "MTH.200.C", "MTH.100.C", 
+                                   "Bitumen", "Lubricants", "Other oil products", 
+                                   "LTH.20.C", 
+                                   "Hard coal (if no detail) [of Coal mines]", 
+                                   "Other bituminous coal [of Coal mines]", 
+                                   "Paraffin waxes", "White spirit & SBP"))
+  # Check the class of the year columns. They should be numeric.
+  expect_true(is.numeric(Eta_fu_template_2019[["1971"]]))
+  expect_true(is.numeric(Eta_fu_template_2019[["2000"]]))
+})
+
+
 test_that("write_eta_fu_template works as expected for 2018 data", {
   # Try with default sort order
   Eta_fu_template_2018 <- load_fu_allocation_data(sample_fu_allocation_table_path(2018)) %>% 
@@ -348,7 +376,7 @@ test_that("write_eta_fu_template works as expected for 2019 data", {
 })
 
 
-test_that("load_eta_fu_data works as expected", {
+test_that("load_eta_fu_data() works as expected", {
   # Load eta_fu data from package
   for (year in valid_iea_release_years) {
     Eta_fu <- load_eta_fu_data(path = sample_eta_fu_table_path(year))
@@ -358,4 +386,20 @@ test_that("load_eta_fu_data works as expected", {
 })
 
 
+test_that("check_fu_allocation_data() works as expected", {
+  expect_true(load_fu_allocation_data() %>% check_fu_allocation_data())
+  
+  # Make a bogus fu_allocation data frame that should fail and make sure it fails
+  # in the situation where Ef.product and Eu.product are not same when Machine is Non-energy.
+  fu_allocation_bad <- tibble::tribble(~Country, ~Flow.aggregation.point, ~Ef.product, ~Machine, ~Eu.product, ~Destination, 
+                                       "Wakanda", "Consumption", "Bitumen", "Non-energy", "Bituman", "Road")
+  expect_error(check_fu_allocation_data(fu_allocation_bad), 
+               "Ef.product and Eu.product must be identical when Machine is Non-energy. The following combinations do not meet that criterion: Wakanda, Consumption, Bitumen, Non-energy, Bituman, Road. Please check the FU allocation table for typos or misspellings.")
+  
+  # Make a bogus fu_allocation data frame that should fail and make sure it fails
+  # in the situation where .values is not NA.
+  fu_allocation_bad2 <- tibble::tribble(~Country, ~Year, ~Flow.aggregation.point, ~Ef.product, ~Machine, ~Eu.product, ~Destination, ~Quantity, ~.values,
+                                        "Wakanda", 2020, "Consumption", "Electricity", NA_character_, "MD", "Industry", "C_1 [%]", "25.0")
+  expect_error(check_fu_allocation_data(fu_allocation_bad2), "In the FU Allocations tab, Eu.product and Destination must be filled when Quantity is non-zero. The following combinations do not meet that criterion: Wakanda, 2020, Consumption, Electricity, NA, MD, Industry, C_1")
+})